Added products check on category delete

+++ Schema/Category/mutation.js
@@ -1,4 +1,5 @@
 import { CategoryModel } from "./db.js";
+import { ProductModel } from "../Products/db.js";
 
 export const CategoryMutation = {
   categoryCreate: async (_, args) => {
@@ -17,6 +18,10 @@ export const CategoryMutation = {
 
   categoryDelete: async (_, args) => {
     const { id } = args;
+    const products = await ProductModel.find({ category: id });
+    if (products.length > 0) {
+      throw new Error("Category cannot be deleted, products exist");
+    }
     await CategoryModel.findByIdAndDelete(id);
     return "Category deleted successfully";
   },
